requirements.txt
# --- Core Web Framework ---
flask
flask_sqlalchemy
flask_migrate
werkzeug
python-dotenv

# --- Data Processing & Statistics ---
pandas
numpy
scipy
statsmodels
linearmodels
pmdarima
scikit-learn

# --- File Handling (Excel Support) ---
openpyxl

# --- Reporting (PDF Generation) ---
fpdf

# --- Image Processing & AI (Updated) ---
Pillow
pytesseract
opencv-python
google-generativeai 

# --- Visualization ---
matplotlib
seaborn